Update submodules to use absolute paths.
Currently the paths for the submodules are relative urls. This means if
the original repo is https://github.com/JeffersonLab/chroma.git then
git will try to pull other_libs/cpp_wilson_dslash from
https://github.com/JeffersonLab/chroma.git/../cpp_wilson_dslash.git ->
https://github.com/JeffersonLab/cpp_wilson_dslash.git

But once the repo has been forked, then the source repo becomes (e.g.)
https://github.com/rabbott999/chroma.git so git will try to look at
https://github.com/rabbott999/cpp_wilson_dslash.git, which does not
exist unless one also forks all of the submodules.

There are 2 possible solutions here as far as I can see:

(1) Fork all of the submodules in addition to chroma

(2) Change the relative paths to absolute paths, so git always looks at
the JLab repo

I've done (2) here. It would be possible to do (1), but as long as I'm
not modifying those repos then it doesn't really make sense to maintain
a copy just so that git finds it in the right place.